requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.13.1
beautifulsoup4==4.12.2
bertviz==1.4.0
biopython==1.81
black==24.2.0
bleach==6.1.0
boto3==1.34.51
botocore==1.34.51
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.2.0
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.4
defusedxml==0.7.1
dill==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.6.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.3
fastjsonschema==2.19.0
filelock==3.11.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.2.0
gitdb==4.0.10
GitPython==3.1.31
hjson==3.1.0
huggingface-hub==0.21.1
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.1
llvmlite==0.42.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
metric-learn==0.6.2
mistune==3.0.2
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
overrides==7.4.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
pox==0.3.2
ppft==1.7.6.6
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.19.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.11.1
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.0
transformers==4.38.1
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
wandb==0.14.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
yarl==1.8.2
zipp==3.15.0
